All buttons are working
Tab switching is not working yet.
I have set up the addTask function which will take the data from the inputs, convert into into an object
and then send it off to whereever it may need to go. In this case, it will be going to eventListeners.js so that when the button is submit button is clicked, the data will be retrieved then we can send that data off to a new module that will create the DOM.